pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "email-validator>=2.2.0",
    "flask>=3.0.3",
    "flask-sqlalchemy>=3.1.1",
    "psycopg2-binary>=2.9.10",
    "nltk>=3.9.1",
    "spacy>=3.8.2",
    "textblob>=0.18.0.post0",
    "scikit-learn>=1.5.2",
    "google-generativeai>=0.8.3",
]
